+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log
 
+## 0.21.0 - 2020-03-17
+
+- Add "Newest projects" section to the top page, and sort by date of addition
+- Add "Featured Projects" section to the top page
+- Add story and rankings frm the weekly newsletter to the top page
+
 ## 0.20.0 - 2019-12-26
 
 - Add monthly trends graph to project details page (#81)